lst/requirements.txt
asgiref==3.8.1
config==0.5.1
Django==5.0.4
django-environ==0.11.2
psycopg2==2.9.9
PyJWT==2.8.0
sqlparse==0.5.0
